backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3&gt;=1.34.129
requests-oauthlib&gt;=2.0.0
cryptography&gt;=42.0.8
python-dotenv&gt;=1.0.1
pymongo==4.5.0
pydantic&gt;=2.6.4
email-validator&gt;=2.2.0
pyjwt&gt;=2.10.1
passlib&gt;=1.7.4
tzdata&gt;=2024.2
motor==3.3.1
pytest&gt;=8.0.0
black&gt;=24.1.1
isort&gt;=5.13.2
flake8&gt;=7.0.0
mypy&gt;=1.8.0
python-jose&gt;=3.3.0
requests&gt;=2.31.0
pandas&gt;=2.2.0
numpy&gt;=1.26.0
python-multipart&gt;=0.0.9
jq&gt;=1.6.0
typer&gt;=0.9.0
reportlab&gt;=4.2.2